GuiImpl::RunMsg: iterate portrait/dialogue case codegen (77%)
- portrait/showPortrait args use the right field widths (i16 vs i32 dword)
- portrait cases use if/else instead of ternary (forces branch, not setcc)
- case 15/17 read the portrait index as the packed dword (ZUN quirk)
RunMsg fn_diff 67% -> 77%. Remaining ~23% is per-case codegen iteration.

+++ src/Gui.hpp
@@ -61,11 +61,11 @@ union MsgRawInstrArgs {
     } portrait;
     struct
     {
-        i16 portraitIdx;
-        i16 anmScriptIdx0;
-        i16 anmScriptIdx1;
-        i16 anmScriptIdx2;
-        i16 anmScriptIdx3;
+        i32 portraitIdx;
+        i32 anmScriptIdx0;
+        i32 anmScriptIdx1;
+        i32 anmScriptIdx2;
+        i32 anmScriptIdx3;
     } showPortrait;
     struct
     {
